requirements.txt
discord.py==1.3.4
# https://github.com/Rapptz/discord.py/archive/rewrite.zip
